migrate most to new DataType
Move original model package to modelORIG and rename model2 to model.
Add deprecated methods to get things to compile.
Still need to deal with Time.
