+++ core/src/main/scala/latis/data/LinearSet1D.scala
@@ -1,7 +1,7 @@
 package latis.data
 
-import latis.metadata._
 import latis.model._
+import latis.util.Identifier.IdentifierStringContext
 import latis.util.LatisException
 
 /**
@@ -20,8 +20,7 @@ class LinearSet1D(start: Double, increment: Double, count: Int)
    * Define the model of this DomainSet assuming double value types
    * and a 1-based "_#" naming scheme for variable identifiers.
    */
-  def model: DataType =
-    Scalar(Metadata("id" -> "_1", "type" -> "double"))
+  def model: DataType = Scalar(id"_1", DoubleValueType)
 
   /**
    * Return the number of values in this DomainSet.
