+++ core/src/main/scala/latis/model/Index.scala
@@ -1,28 +1,10 @@
 package latis.model
 
-import latis.metadata.Metadata
+import latis.model.LongValueType
 import latis.util.Identifier
 
-/**
- * Index is a Scalar that is used as a placeholder for a domain variable
- * that does not have a value represented in a Sample.
- */
-class Index private (metadata: Metadata) extends Scalar(metadata) {
-  //TODO: enforce that Index variables only appear in Function domains
-  //TODO: make sure operations handle Index appropriately
-
-  /** Override to preserve type */
-  override def rename(id: Identifier): Index = Index(id)
-}
+class Index(id: Identifier) extends Scalar(id, LongValueType)
 
 object Index {
-
-  /**
-   * Constructs an Index variable with the given id.
-   */
-  def apply(id: Identifier): Index =
-    new Index(Metadata(
-      "id" -> id.asString,
-      "type" -> "long" //Scalar needs value type
-    ))
+  def apply(id: Identifier): Index = new Index(id)
 }
