--- core/src/main/scala/latis/model2/Index.scala
@@ -1,10 +1,0 @@
-package latis.model2
-
-import latis.model.LongValueType
-import latis.util.Identifier
-
-class Index(id: Identifier) extends Scalar(id, LongValueType)
-
-object Index {
-  def apply(id: Identifier): Index = new Index(id)
-}
