+++ core/src/main/scala/latis/ops/Curry.scala
@@ -40,7 +40,7 @@ case class Curry(arity: Int = 1) extends GroupOperation {
     case Function(d, _) =>
       d.getScalars.take(arity) match { //TODO: fails for nested Tuples in domain
         case s1 :: Nil => s1
-        case ss => Tuple(ss)
+        case ss => Tuple.fromSeq(ss).fold(throw _, identity)
       }
     //case Function(Tuple(es @ _*), _) => Tuple(es.take(arity))
     case _ => throw LatisException("Invalid DataType for Curry")
@@ -57,8 +57,8 @@ case class Curry(arity: Int = 1) extends GroupOperation {
         case Function(d, range) =>
           (d.getScalars.drop(arity) match { //TODO: fails for nested Tuples in domain
             case Nil => range  // this happens when the arity is not changed
-            case s1 :: Nil => Function(s1, range)
-            case ss => Function(Tuple(ss), range)
+            case s1 :: Nil => Function.from(s1, range).fold(throw _, identity)
+            case ss => Function.from(Tuple.fromSeq(ss).fold(throw _, identity), range).fold(throw _, identity)
           }).asRight
         case _ => Left(LatisException("Model must be a function"))
         //case Function(Tuple(es @ _*), range) => Function(Tuple(es.drop(arity)), range)
