+++ core/src/main/scala/latis/ops/GroupByVariable.scala
@@ -50,7 +50,7 @@ case class GroupByVariable(ids: Identifier*) extends GroupOperation {
           throw LatisException(msg)
       }
     }
-    Tuple(scalars).flatten
+    Tuple.fromSeq(scalars).fold(throw _, identity)  //TODO: .flatten?
   }
 
   def groupByFunction(model: DataType): Sample => Option[DomainData] =
@@ -77,17 +77,17 @@ case class RemoveGroupedVariables(ids: Seq[Identifier]) extends MapOperation {
   /** Recursive method to build new model by dropping variableNames. */
   private def applyToVariable(v: DataType): Option[DataType] = v match {
     case s: Scalar =>
-      if (s.id.exists(id => ids.contains(id))) None else Some(s)
+      if (ids.contains(s.id)) None else Some(s)
     case t @ Tuple(vars @ _*) =>
       val vs = vars.flatMap(applyToVariable)
       vs.length match {
         case 0 => None // drop empty Tuple
         case 1 => Some(vs.head) // reduce Tuple of one
-        case _ => Some(Tuple(t.metadata, vs))
+        case _ => Some(Tuple.fromSeq(t.id, vs).fold(throw _, identity))
       }
     case f @ Function(d, r) =>
       (applyToVariable(d), applyToVariable(r)) match {
-        case (Some(d), Some(r)) => Some(Function(f.metadata, d, r))
+        case (Some(d), Some(r)) => Some(Function.from(f.id, d, r).fold(throw _, identity))
         case (None, Some(r)) => Some(r)
         //TODO: deal with empty range
         case _ => None
@@ -96,9 +96,7 @@ case class RemoveGroupedVariables(ids: Seq[Identifier]) extends MapOperation {
 
   override def mapFunction(model: DataType): Sample => Sample = {
     // Determine the list of variables to keep
-    val vnames: List[Identifier] = model.getScalars.map(
-      _.id.getOrElse(throw LatisException("Found an unnamed Scalar"))
-    ).filterNot(ids.contains)
+    val vnames: List[Identifier] = model.getScalars.map(_.id).filterNot(ids.contains)
 
     // Get the paths of the variables to be removed from each Sample.
     // Sort to maintain the original order of variables.
