+++ core/src/main/scala/latis/ops/TimeOperation.scala
@@ -37,8 +37,8 @@ abstract class TimeOperation extends MapOperation {
    */
   private def makeConverters(model: DataType): List[(SamplePosition, Datum => Datum)] =
     model.getScalars.collect {
-      case t: Time if (t.id.nonEmpty) =>  //TODO: require scalars to have ID
-        val path = model.getPath(t.id.get).get //bug if path not found
+      case t: Time =>
+        val path = model.getPath(t.id).get //bug if path not found
         val pos = path.length match {
           case 0 => RangePosition(0) //hack for ConstantFunction
           case 1 => path.head
