+++ core/src/test/scala/latis/model/FunctionSuite.scala
@@ -1,4 +1,4 @@
-package latis.model2
+package latis.model
 
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.Inside.inside
