+++ core/src/test/scala/latis/modelORIG/FillValueSuite.scala
@@ -1,4 +1,4 @@
-package latis.model
+package latis.modelORIG
 
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.Inside.inside
